# backend/openapi.yaml
openapi: 3.0.3
info:
  title: V PATHing Rewards API
  version: 1.0.0
  description: Cashback rewards platform API
servers:
  - url: https://vpathrewards.store/api
    description: Production
security:
  - bearerAuth: []
  - cookieAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Goal:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        title:
          type: string
        target_amount:
          type: number
        period_type:
          type: string
          enum: [monthly, weekly, yearly, custom]
        start_date:
          type: string
          format: date
          nullable: true
        end_date:
          type: string
          format: date
          nullable: true
...
paths:
...
  /cashback/transactions:
    get:
      operationId: getCashbackTransactions
      summary: Get cashback transactions
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Transactions
...
  /cashback/history:
    get:
      operationId: getCashbackHistory
      summary: Get cashback history
      parameters:
        - name: group_by
          in: query
          schema:
            type: string
            enum: [day, week, month]
        - name: days
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 365
        - name: status
          in: query
          schema:
            type: string
      responses:
        '200':
          description: History
  /cashback/calendar:
    get:
      operationId: getCashbackCalendar
      summary: Get cashback calendar
      parameters:
        - name: year
          in: query
          schema:
            type: integer
            minimum: 2020
            maximum: 2100
        - name: month
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 12
      responses:
        '200':
          description: Calendar data
  /cashback/goals:
    get:
      operationId: getGoals
      summary: Get cashback goals
      responses:
        '200':
          description: Goals

    post:
      operationId: createGoal
      summary: Create a cashback goal
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, target_amount]
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                target_amount:
                  type: number
                  minimum: 0.01
                period_type:
                  type: string
                  enum: [monthly, weekly, yearly, custom]
                  default: monthly
                start_date:
                  type: string
                  format: date
                  nullable: true
                end_date:
                  type: string
                  format: date
                  nullable: true
      responses:
        '201':
          description: Goal created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
  /cashback/goals/{id}:
    put:
      operationId: updateGoal
      summary: Update a cashback goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 1
                  maxLength: 200
                target_amount:
                  type: number
                  minimum: 0.01
                period_type:
                  type: string
                  enum: [monthly, weekly, yearly, custom]
                start_date:
                  type: string
                  format: date
                  nullable: true
                end_date:
                  type: string
                  format: date
                  nullable: true
      responses:
        '200':
          description: Goal updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Goal'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: deleteGoal
      summary: Delete a cashback goal
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        '200':
          description: Goal deleted
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /cashback/track:
    post:
      operationId: trackCashback
      summary: Track a cashback transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [offer_id, amount]
              properties:
                offer_id:
                  type: integer
                  minimum: 1
                amount:
                  type: number
                  minimum: 0.01
      responses:
        '200':
          description: Tracked
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ─── Search ────────────────────────────────────────────────────────────────
...